requirements.txt
# pip install -r requirements.txt

flask
requests
Pillow